requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
unidecode>=1.3.0
optimizely-opal-opal-tools-sdk
httpx>=0.25.0
webtech
gunicorn